Move isolation to worker module
